pyproject.toml
[project]
name = "ideeza-analytics-api"
version = "0.1.0"
description = "IDEEZA Analytics API - Backend assessment project"
requires-python = ">=3.12"
dependencies = [
    "Django==5.2.9",
    "django-environ==0.11.2",
    "psycopg2-binary==2.9.9",
    "djangorestframework==3.16.1",
    "drf-yasg=1.21.11",
    "django-filter==25.2",
    "whitenoise==6.11.0",
    "django-cors-headers==4.9.0",
    "django-extensions==4.1",
    "gunicorn==23.0.0",
]

[project.optional-dependencies]
dev = [
    "django-debug-toolbar==6.1.0",
    "mypy==1.18.2",
    "django-stubs==5.0.0",
    "ruff==0.14.5",
    "pre-commit==4.5.0",
    "rich==14.1.0",
    "pytest==8.3.3",
    "pytest-django==4.8.0",
    "pytest-cov==5.0.0",
    "pytest-mock==3.14.0",
    "factory-boy==3.3.0",
    "faker==24.4.0",
    "pytest-xdist==3.5.0",
    "pytest-timeout==2.2.0",
]

# General Settings
[tool.ruff]
line-length = 120
indent-width = 4
preview = true
target-version = "py312"
output-format = "grouped"
respect-gitignore = true
show-fixes = true
fix = true
unsafe-fixes = true
exclude = [ 
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "**/migrations/**",
    "node_modules",
    "site-packages",
    ".venv",
    "venv",
    "codegen",
]

[tool.ruff.lint]
select = ["COM", "DJ", "E", "F", "I", "LOG", "N", "PIE", "Q", "RET", "T10", "T20", "W", "I001"]
ignore = ["DJ001", "E722", "COM812"]
fixable = ["ALL"]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Per-file ignores
[tool.ruff.lint.per-file-ignores]
"**/migrations/**/*.py" = ["E501"]  # E501: Allow long lines in migration files (auto-generated)

# Formatting Settings
[tool.ruff.format]
indent-style = "space"
line-ending = "auto"
quote-style = "double"
skip-magic-trailing-comma = false

# Pydocstyle Settings
[tool.ruff.lint.pydocstyle]
convention = "google"

# Isort Settings
[tool.ruff.lint.isort]
case-sensitive = true
known-first-party = ["core", "config"]
known-third-party = ["django", "rest_framework", "drf_yasg"]

# Mypy Settings
[tool.mypy]
plugins = []

# Basedpyright Settings
[tool.basedpyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
extraPaths = [".venv/lib/python3.12/site-packages"]
reportMissingImports = "error"
useLibraryCodeForTypes = true
exclude = [
    "**/node_modules",
    "**/__pycache__",
    "**/.*",
    ".venv",
    "venv",
    "migrations",
    "codegen",
]

[dependency-groups]
dev = [
    "ruff==0.14.5",
]
